requirements.txt
Flask
gunicorn
numpy
scikit-learn
pandas
PyYAML
